refactor(suggest): Split merino-suggest into a -traits and a -providers
Moving the provider implementations away from the basic traits makes room to have tools like `make_provider_tree` in a common place without creating dependency loops.
