+++ Cargo.toml
@@ -6,7 +6,8 @@ members = [
   "merino-cache",
   "merino-integration-tests-macro",
   "merino-integration-tests",
-  "merino-suggest",
+  "merino-suggest-traits",
+  "merino-suggest-providers",
   "merino-web",
   "merino-wikipedia",
   "merino",
