+++ merino-adm/Cargo.toml
@@ -12,15 +12,13 @@ futures = "0.3.14"
 http = "0.2.4"
 lazy_static = "1.4.0"
 merino-settings = { path = "../merino-settings" }
-merino-suggest = { path = "../merino-suggest" }
-radix_trie = "0.2.1"
+merino-suggest-traits = { path = "../merino-suggest-traits" }
 reqwest = { version = "0.11.3", features = ["json"] }
 # Pin sentry_anyhow to 0.19 until our on-premise server updates to 20.6.
 sentry-anyhow = "0.19"
 serde = { version = "1.0.125", features = ["rc"] }
 serde_derive = "1.0.125"
 serde_json = "1.0.64"
-serde_qs = "0.8.3"
 serde_with = "1.9"
 thiserror = "1.0"
 tokio = { version = "1.8.2", features = ["rt", "macros", "rt-multi-thread"] }
@@ -36,5 +34,3 @@ features = ["backtrace", "contexts", "panic", "transport", "anyhow"]
 [dev-dependencies]
 actix-rt = "2.2"
 fake = "2"
-httpmock = "0.5.8"
-pretty_assertions = "0.7.1"
