+++ merino-cache/Cargo.toml
@@ -13,7 +13,7 @@ deduped-dashmap = { path = "../deduped-dashmap" }
 fix-hidden-lifetime-bug = "0.2.4"
 lazy_static = "1.4"
 merino-settings = { path = "../merino-settings" }
-merino-suggest = { path = "../merino-suggest" }
+merino-suggest-traits = { path = "../merino-suggest-traits" }
 redis = { version = "^0.20", features = ["tokio-comp", "connection-manager"] }
 serde = "^1"
 serde_json = "^1"
