+++ merino-integration-tests-macro/Cargo.toml
@@ -10,7 +10,3 @@ proc-macro = true
 proc-macro2 = "1"
 quote = "1"
 syn = { version = "^1", features = ["full"] }
-
-[dev-dependencies]
-merino-integration-tests = { path = "../merino-integration-tests" }
-parameterized = "0.3"
