+++ merino-integration-tests/Cargo.toml
@@ -9,9 +9,7 @@ anyhow = "1.0.40"
 cadence = "0.26"
 crossbeam-channel = "0.5"
 deadpool = "^0.8"
-httpmock = "0.5.8"
 lazy_static = "^1.4.0"
-maplit = "1.0.2"
 merino-integration-tests-macro = { path = "../merino-integration-tests-macro" }
 merino-settings = { path = "../merino-settings" }
 merino-web = { path = "../merino-web" }
@@ -24,7 +22,8 @@ serde_json = "1.0.64"
 serde_with = "1.8.1"
 statsd-parser = "0.3"
 tokio = "1.8.2"
-tokio-test = "0.4.1"
 tracing = "0.1.29"
 tracing-futures = "0.2"
 tracing-subscriber = "0.2.18"
+
+[features]
