+++ merino-suggest-providers/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "merino-suggest-providers"
+version = "0.5.1"
+edition = "2021"
+
+[dependencies]
+actix-web = { version = "=4.0.0-beta.15", default_features = false }
+anyhow = "1.0"
+async-trait = "0.1"
+blake3 = "1"
+cadence = "0.26"
+futures = "0.3"
+http = "0.2.4"
+merino-adm = { path = "../merino-adm" }
+merino-cache = { path = "../merino-cache" }
+merino-settings = { path = "../merino-settings" }
+merino-suggest-traits = { path = "../merino-suggest-traits" }
+serde = { version = "1.0.125", features = ["derive"] }
+serde_json = "1.0"
+serde_with = "1.9.1"
+thiserror = "1.0"
+fake = { version = "2.4", features = ["derive"] }
+rand = "0.8"
+tokio = { version = "1", features = ["time"] }
+uuid = "0.8.2"
+regex = "1.5.4"
+async-recursion = "1.0.0"
+
+[dev_dependencies]
+tokio = { version = "1", features = ["time", "macros", "rt", "sync"] }
