+++ merino-suggest-providers/src/providers/mod.rs
@@ -1,0 +1,11 @@
+//! Merino providers that aren't tied to any particular data source.
+
+pub(crate) mod client_variant_filter;
+pub(crate) mod debug;
+pub(crate) mod fixed;
+pub(crate) mod id_multi;
+pub(crate) mod keyword_filter;
+pub(crate) mod multi;
+pub(crate) mod stealth;
+pub(crate) mod timeout;
+pub(crate) mod wikifruit;
