+++ merino-suggest-traits/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "merino-suggest"
+name = "merino-suggest-traits"
 version = "0.5.1"
 edition = "2021"
 
@@ -9,9 +9,7 @@ anyhow = "1.0"
 async-trait = "0.1"
 blake3 = "1"
 cadence = "0.26"
-futures = "0.3"
 http = "0.2.4"
-merino-settings = { path = "../merino-settings" }
 serde = { version = "1.0.125", features = ["derive"] }
 serde_json = "1.0"
 serde_with = "1.9.1"
@@ -19,8 +17,9 @@ thiserror = "1.0"
 fake = { version = "2.4", features = ["derive"] }
 rand = "0.8"
 tokio = { version = "1", features = ["time"] }
-uuid = "0.8.2"
 regex = "1.5.4"
 
 [dev_dependencies]
 tokio = { version = "1", features = ["time", "macros", "rt", "sync"] }
+
+[features]
