+++ merino-suggest-traits/src/device_info.rs
