+++ merino-web/Cargo.toml
@@ -17,7 +17,8 @@ lazy_static = "1.4.0"
 merino-adm = { path = "../merino-adm" }
 merino-cache = { path = "../merino-cache" }
 merino-settings = { path = "../merino-settings" }
-merino-suggest = { path = "../merino-suggest" }
+merino-suggest-providers = { path = "../merino-suggest-providers" }
+merino-suggest-traits = { path = "../merino-suggest-traits" }
 # Pin sentry_backtrace to 0.19 until our on-premise server updates to 20.6.
 sentry-backtrace = "0.19"
 serde = { version = "1.0.125", features = ["derive"] }
@@ -25,7 +26,6 @@ serde_json = "1.0.64"
 serde_with = "1.9"
 thiserror = "1.0.24"
 tokio = { version = "1.8.2", features = ["sync"] }
-tokio-test = "0.4.1"
 tracing = { version = "0.1.29", features = ["async-await"] }
 tracing-actix-web-mozlog = "0.4.1"
 tracing-futures = "0.2"
@@ -53,3 +53,4 @@ features = ["backtrace", "contexts", "panic", "transport", "anyhow"]
 [dev-dependencies]
 actix-rt = "2.2.0"
 pretty_assertions = "0.7"
+tokio-test = "0.4.2"
